Created a super/subclass structure to JSONScanner to reduce the complexity of the JSONScanner class
