+++ src/main/java/com/alibaba/fastjson/parser/scanner/JSONScanner.java
@@ -1,0 +1,198 @@
+/*
+ * Copyright 1999-2101 Alibaba Group.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.alibaba.fastjson.parser.scanner;
+
+import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.parser.JSONLexer;
+import com.alibaba.fastjson.parser.JSONToken;
+import com.alibaba.fastjson.parser.dtos.AddSymbolDTO;
+import com.alibaba.fastjson.parser.dtos.ArrayCopyDTO;
+import com.alibaba.fastjson.util.Base64;
+
+//这个类，为了性能优化做了很多特别处理，一切都是为了性能！！！
+
+/**
+ * @author wenshao<szujobs@hotmail.com>
+ */
+public class JSONScanner extends JSONLexer {
+
+    private String text;
+
+    public JSONScanner(String input){
+        this(input, JSON.DEFAULT_PARSER_FEATURE);
+    }
+
+    public JSONScanner(String input, int features){
+        this.features = features;
+
+        text = input;
+        bp = -1;
+
+        next();
+        if (ch == 65279) {
+            next();
+        }
+    }
+
+    public char charAt(int index) {
+        if (index >= text.length()) {
+            return EOI;
+        }
+
+        return text.charAt(index);
+    }
+
+    public char next() {
+        return ch = charAt(++bp);
+    }
+
+    public JSONScanner(char[] input, int inputLength){
+        this(input, inputLength, JSON.DEFAULT_PARSER_FEATURE);
+    }
+
+    public JSONScanner(char[] input, int inputLength, int features){
+        this(new String(input, 0, inputLength), features);
+    }
+
+    protected final void copyTo(int offset, int count, char[] dest) {
+        text.getChars(offset, offset + count, dest, 0);
+    }
+
+    protected final static char[] typeFieldName = ("\"" + JSON.DEFAULT_TYPE_KEY + "\":\"").toCharArray();
+
+    public final int scanType(String type) {
+        matchStat = UNKOWN;
+
+        if (!charArrayCompare(text, bp, typeFieldName)) {
+            return NOT_MATCH_NAME;
+        }
+
+        int bpLocal = this.bp + typeFieldName.length;
+
+        final int typeLength = type.length();
+        for (int i = 0; i < typeLength; ++i) {
+            if (type.charAt(i) != charAt(bpLocal + i)) {
+                return NOT_MATCH;
+            }
+        }
+        bpLocal += typeLength;
+        if (charAt(bpLocal) != '"') {
+            return NOT_MATCH;
+        }
+
+        this.ch = charAt(++bpLocal);
+
+        if (ch == ',') {
+            this.ch = charAt(++bpLocal);
+            this.bp = bpLocal;
+            token = JSONToken.COMMA;
+            return VALUE;
+        } else if (ch == '}') {
+            ch = charAt(++bpLocal);
+            if (ch == ',') {
+                token = JSONToken.COMMA;
+                this.ch = charAt(++bpLocal);
+            } else if (ch == ']') {
+                token = JSONToken.RBRACKET;
+                this.ch = charAt(++bpLocal);
+            } else if (ch == '}') {
+                token = JSONToken.RBRACE;
+                this.ch = charAt(++bpLocal);
+            } else if (ch == EOI) {
+                token = JSONToken.EOF;
+            } else {
+                return NOT_MATCH;
+            }
+            matchStat = END;
+        }
+
+        this.bp = bpLocal;
+        return matchStat;
+    }
+
+    static final boolean charArrayCompare(String src, int offset, char[] dest) {
+        final int destLen = dest.length;
+        if (destLen + offset > src.length()) {
+            return false;
+        }
+
+        for (int i = 0; i < destLen; ++i) {
+            if (dest[i] != src.charAt(offset + i)) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    public final boolean charArrayCompare(char[] chars) {
+        return charArrayCompare(text, bp, chars);
+    }
+
+    public final int indexOf(char ch, int startIndex) {
+        return text.indexOf(ch, startIndex);
+    }
+
+    public final String addSymbol(AddSymbolDTO parameterObject) {
+        return parameterObject.symbolTable.addSymbol(text, parameterObject.offset, parameterObject.len, parameterObject.hash);
+    }
+
+    public byte[] bytesValue() {
+        return Base64.decodeFast(text, np + 1, sp);
+    }
+
+    /**
+     * The value of a literal token, recorded as a string. For integers, leading 0x and 'l' suffixes are suppressed.
+     */
+    public final String stringVal() {
+        if (!hasSpecial) {
+            // return new String(buf, np + 1, sp);
+            return text.substring(np + 1, np + 1 + sp);
+        } else {
+            return new String(sbuf, 0, sp);
+        }
+    }
+
+    public final String subString(int offset, int count) {
+        return text.substring(offset, offset + count);
+    }
+
+    public final String numberString() {
+        char chLocal = charAt(np + sp - 1);
+
+        int sp = this.sp;
+        if (chLocal == 'L' || chLocal == 'S' || chLocal == 'B' || chLocal == 'F' || chLocal == 'D') {
+            sp--;
+        }
+
+        return text.substring(np, np + sp);
+        // return new String(buf, np, sp);
+    }
+
+    public final int ISO8601_LEN_0 = "0000-00-00".length();
+    public final int ISO8601_LEN_1 = "0000-00-00T00:00:00".length();
+    public final int ISO8601_LEN_2 = "0000-00-00T00:00:00.000".length();
+
+    @Override
+    public boolean isEOF() {
+        return bp == text.length() || ch == EOI && bp + 1 == text.length();
+    }
+
+
+    protected final void arrayCopy(ArrayCopyDTO parameterObject) {
+        text.getChars(parameterObject.srcPos, parameterObject.srcPos + parameterObject.length, parameterObject.dest, parameterObject.destPos);
+    }
+}
