+++ src/main/java/com/alibaba/fastjson/parser/scanner/JSONScannerDateTime.java
@@ -1,0 +1,354 @@
+package com.alibaba.fastjson.parser.scanner;
+
+import java.util.Calendar;
+import java.util.Locale;
+import java.util.TimeZone;
+
+import com.alibaba.fastjson.parser.JSONToken;
+
+public class JSONScannerDateTime extends JSONScanner{
+	private String text;
+	 
+	public JSONScannerDateTime(String input){
+       super(input);
+    }
+
+    public JSONScannerDateTime(char[] input, int inputLength){
+        super(input, inputLength);
+    }
+
+    public JSONScannerDateTime(char[] input, int inputLength, int features){
+        super(input, inputLength, features);
+    }
+	
+	public JSONScannerDateTime(String input, int features) {
+		super(input, features);
+	}
+	
+
+    public boolean scanISO8601DateIfMatch(boolean strict) {
+        int rest = text.length() - bp;
+
+        if ((!strict) && rest > 13) {
+            char c0 = charAt(bp);
+            char c1 = charAt(bp + 1);
+            char c2 = charAt(bp + 2);
+            char c3 = charAt(bp + 3);
+            char c4 = charAt(bp + 4);
+            char c5 = charAt(bp + 5);
+
+            char c_r0 = charAt(bp + rest - 1);
+            char c_r1 = charAt(bp + rest - 2);
+            if (c0 == '/' && c1 == 'D' && c2 == 'a' && c3 == 't' && c4 == 'e' && c5 == '(' && c_r0 == '/'
+                && c_r1 == ')') {
+                int plusIndex = -1;
+                for (int i = 6; i < rest; ++i) {
+                    char c = charAt(bp + i);
+                    if (c == '+') {
+                        plusIndex = i;
+                    } else if (c < '0' || c > '9') {
+                        break;
+                    }
+                }
+                if (plusIndex == -1) {
+                    return false;
+                }
+                int offset = bp + 6;
+                String numberText = this.subString(offset, plusIndex - offset);
+                long millis = Long.parseLong(numberText);
+
+                Locale local = Locale.getDefault();
+                calendar = Calendar.getInstance(TimeZone.getDefault(), local);
+                calendar.setTimeInMillis(millis);
+
+                token = JSONToken.LITERAL_ISO8601_DATE;
+                return true;
+            }
+        }
+
+        if (rest == 8 || rest == 14 || rest == 17) {
+            if (strict) {
+                return false;
+            }
+
+            char y0 = charAt(bp);
+            char y1 = charAt(bp + 1);
+            char y2 = charAt(bp + 2);
+            char y3 = charAt(bp + 3);
+            char M0 = charAt(bp + 4);
+            char M1 = charAt(bp + 5);
+            char d0 = charAt(bp + 6);
+            char d1 = charAt(bp + 7);
+
+            if (!checkDate(y0, y1, y2, y3, M0, M1, d0, d1)) {
+                return false;
+            }
+
+            setCalendar(y0, y1, y2, y3, M0, M1, d0, d1);
+
+            int hour, minute, seconds, millis;
+            if (rest != 8) {
+                char h0 = charAt(bp + 8);
+                char h1 = charAt(bp + 9);
+                char m0 = charAt(bp + 10);
+                char m1 = charAt(bp + 11);
+                char s0 = charAt(bp + 12);
+                char s1 = charAt(bp + 13);
+
+                if (!checkTime(h0, h1, m0, m1, s0, s1)) {
+                    return false;
+                }
+
+                if (rest == 17) {
+                    char S0 = charAt(bp + 14);
+                    char S1 = charAt(bp + 15);
+                    char S2 = charAt(bp + 16);
+                    if (S0 < '0' || S0 > '9') {
+                        return false;
+                    }
+                    if (S1 < '0' || S1 > '9') {
+                        return false;
+                    }
+                    if (S2 < '0' || S2 > '9') {
+                        return false;
+                    }
+
+                    millis = digits[S0] * 100 + digits[S1] * 10 + digits[S2];
+                } else {
+                    millis = 0;
+                }
+
+                hour = digits[h0] * 10 + digits[h1];
+                minute = digits[m0] * 10 + digits[m1];
+                seconds = digits[s0] * 10 + digits[s1];
+            } else {
+                hour = 0;
+                minute = 0;
+                seconds = 0;
+                millis = 0;
+            }
+
+            calendar.set(Calendar.HOUR_OF_DAY, hour);
+            calendar.set(Calendar.MINUTE, minute);
+            calendar.set(Calendar.SECOND, seconds);
+            calendar.set(Calendar.MILLISECOND, millis);
+
+            token = JSONToken.LITERAL_ISO8601_DATE;
+            return true;
+        }
+
+        if (rest < ISO8601_LEN_0) {
+            return false;
+        }
+
+        if (charAt(bp + 4) != '-') {
+            return false;
+        }
+        if (charAt(bp + 7) != '-') {
+            return false;
+        }
+
+        char y0 = charAt(bp);
+        char y1 = charAt(bp + 1);
+        char y2 = charAt(bp + 2);
+        char y3 = charAt(bp + 3);
+        char M0 = charAt(bp + 5);
+        char M1 = charAt(bp + 6);
+        char d0 = charAt(bp + 8);
+        char d1 = charAt(bp + 9);
+        if (!checkDate(y0, y1, y2, y3, M0, M1, d0, d1)) {
+            return false;
+        }
+
+        setCalendar(y0, y1, y2, y3, M0, M1, d0, d1);
+
+        char t = charAt(bp + 10);
+        if (t == 'T' || (t == ' ' && !strict)) {
+            if (rest < ISO8601_LEN_1) {
+                return false;
+            }
+        } else if (t == '"' || t == EOI) {
+            calendar.set(Calendar.HOUR_OF_DAY, 0);
+            calendar.set(Calendar.MINUTE, 0);
+            calendar.set(Calendar.SECOND, 0);
+            calendar.set(Calendar.MILLISECOND, 0);
+
+            ch = charAt(bp += 10);
+
+            token = JSONToken.LITERAL_ISO8601_DATE;
+            return true;
+        } else {
+            return false;
+        }
+
+        if (charAt(bp + 13) != ':') {
+            return false;
+        }
+        if (charAt(bp + 16) != ':') {
+            return false;
+        }
+
+        char h0 = charAt(bp + 11);
+        char h1 = charAt(bp + 12);
+        char m0 = charAt(bp + 14);
+        char m1 = charAt(bp + 15);
+        char s0 = charAt(bp + 17);
+        char s1 = charAt(bp + 18);
+
+        if (!checkTime(h0, h1, m0, m1, s0, s1)) {
+            return false;
+        }
+
+        int hour = digits[h0] * 10 + digits[h1];
+        int minute = digits[m0] * 10 + digits[m1];
+        int seconds = digits[s0] * 10 + digits[s1];
+        calendar.set(Calendar.HOUR_OF_DAY, hour);
+        calendar.set(Calendar.MINUTE, minute);
+        calendar.set(Calendar.SECOND, seconds);
+
+        char dot = charAt(bp + 19);
+        if (dot == '.') {
+            if (rest < ISO8601_LEN_2) {
+                return false;
+            }
+        } else {
+            calendar.set(Calendar.MILLISECOND, 0);
+
+            ch = charAt(bp += 19);
+
+            token = JSONToken.LITERAL_ISO8601_DATE;
+            return true;
+        }
+
+        char S0 = charAt(bp + 20);
+        char S1 = charAt(bp + 21);
+        char S2 = charAt(bp + 22);
+        if (S0 < '0' || S0 > '9') {
+            return false;
+        }
+        if (S1 < '0' || S1 > '9') {
+            return false;
+        }
+        if (S2 < '0' || S2 > '9') {
+            return false;
+        }
+
+        int millis = digits[S0] * 100 + digits[S1] * 10 + digits[S2];
+        calendar.set(Calendar.MILLISECOND, millis);
+
+        ch = charAt(bp += 23);
+
+        token = JSONToken.LITERAL_ISO8601_DATE;
+        return true;
+    }
+
+    private boolean checkTime(char h0, char h1, char m0, char m1, char s0, char s1) {
+        if (h0 == '0') {
+            if (h1 < '0' || h1 > '9') {
+                return false;
+            }
+        } else if (h0 == '1') {
+            if (h1 < '0' || h1 > '9') {
+                return false;
+            }
+        } else if (h0 == '2') {
+            if (h1 < '0' || h1 > '4') {
+                return false;
+            }
+        } else {
+            return false;
+        }
+
+        if (m0 >= '0' && m0 <= '5') {
+            if (m1 < '0' || m1 > '9') {
+                return false;
+            }
+        } else if (m0 == '6') {
+            if (m1 != '0') {
+                return false;
+            }
+        } else {
+            return false;
+        }
+
+        if (s0 >= '0' && s0 <= '5') {
+            if (s1 < '0' || s1 > '9') {
+                return false;
+            }
+        } else if (s0 == '6') {
+            if (s1 != '0') {
+                return false;
+            }
+        } else {
+            return false;
+        }
+
+        return true;
+    }
+
+    private void setCalendar(char y0, char y1, char y2, char y3, char M0, char M1, char d0, char d1) {
+        Locale local = Locale.getDefault();
+        calendar = Calendar.getInstance(TimeZone.getDefault(), local);
+        int year = digits[y0] * 1000 + digits[y1] * 100 + digits[y2] * 10 + digits[y3];
+        int month = digits[M0] * 10 + digits[M1] - 1;
+        int day = digits[d0] * 10 + digits[d1];
+        calendar.set(Calendar.YEAR, year);
+        calendar.set(Calendar.MONTH, month);
+        calendar.set(Calendar.DAY_OF_MONTH, day);
+    }
+
+    static boolean checkDate(char y0, char y1, char y2, char y3, char M0, char M1, int d0, int d1) {
+        if (y0 != '1' && y0 != '2') {
+            return false;
+        }
+        if (y1 < '0' || y1 > '9') {
+            return false;
+        }
+        if (y2 < '0' || y2 > '9') {
+            return false;
+        }
+        if (y3 < '0' || y3 > '9') {
+            return false;
+        }
+
+        if (M0 == '0') {
+            if (M1 < '1' || M1 > '9') {
+                return false;
+            }
+        } else if (M0 == '1') {
+            if (M1 != '0' && M1 != '1' && M1 != '2') {
+                return false;
+            }
+        } else {
+            return false;
+        }
+
+        if (d0 == '0') {
+            if (d1 < '1' || d1 > '9') {
+                return false;
+            }
+        } else if (d0 == '1' || d0 == '2') {
+            if (d1 < '0' || d1 > '9') {
+                return false;
+            }
+        } else if (d0 == '3') {
+            if (d1 != '0' && d1 != '1') {
+                return false;
+            }
+        } else {
+            return false;
+        }
+
+        return true;
+    }
+
+	public boolean scanISO8601DateIfMatch() {
+		return scanISO8601DateIfMatch(true);
+	}
+
+
+    
+
+
+
+}
